refactor: lodash imports and _app
