+++ src/components/AdvancedInnerTable/AdvancedInnerTable.tsx
@@ -1,6 +1,6 @@
 import { styled } from '@mui/material';
 import React, { useId } from 'react';
-import { OpenModalTransparency } from '@/views/CoreUnitBudgetStatement/BudgetStatementtUtils';
+import { OpenModalTransparency } from '@/views/CoreUnitBudgetStatement/BudgetStatementUtils';
 import { TransparencyEmptyTable } from '@/views/CoreUnitBudgetStatement/components/Placeholders/TransparencyEmptyTable';
 import { NumberCell } from './NumberCell/NumberCell';
 import { TextCell } from './TextCell/TextCell';
