+++ src/components/BudgetStatement/BudgetStatementForecast/useBudgetStatementForecast.tsx
@@ -1,14 +1,12 @@
 import { styled } from '@mui/material';
 import { API_MONTH_TO_FORMAT } from '@ses/core/utils/date';
 import { capitalizeSentence, getWalletWidthForWallets, toKebabCase } from '@ses/core/utils/string';
-
 import _ from 'lodash';
 import { useRouter } from 'next/router';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import type { InnerTableColumn, InnerTableRow } from '@/components/AdvancedInnerTable/types';
-
 import HeaderWithIcon from '@/components/BudgetStatement/BudgetStatementForecast/HeaderWithIcon/HeaderWithIcon';
-import { renderWallet } from '@/views/CoreUnitBudgetStatement/BudgetStatementtUtils';
+import { renderWallet } from '@/views/CoreUnitBudgetStatement/BudgetStatementUtils';
 import {
   getBudgetCapForMonthOnBudgetStatement,
   getBudgetCapForMonthOnWalletOnBudgetStatement,
@@ -31,8 +29,7 @@ import {
   getForecastBreakdownColumns,
 } from '@/views/CoreUnitBudgetStatement/utils/forecastTableHelpers';
 import replacePaymentTopup from '@/views/CoreUnitBudgetStatement/utils/helpers';
-import ProgressiveIndicator from './ProgresiveIndicator';
-
+import ProgressiveIndicator from './ProgressiveIndicator';
 import type { BudgetStatement } from '@ses/core/models/interfaces/budgetStatement';
 import type { BudgetStatementWallet } from '@ses/core/models/interfaces/budgetStatementWallet';
 import type { DateTime } from 'luxon';
