+++ src/components/BudgetStatement/BudgetStatementTransferRequest/components/TransferRequestSection/TransferRequestSection.stories.tsx
@@ -1,7 +1,7 @@
 import { DateTime } from 'luxon';
 import type { InnerTableColumn, InnerTableRow } from '@/components/AdvancedInnerTable/types';
 import { createThemeModeVariants } from '@/core/utils/storybook/factories';
-import { renderWallet } from '@/views/CoreUnitBudgetStatement/BudgetStatementtUtils';
+import { renderWallet } from '@/views/CoreUnitBudgetStatement/BudgetStatementUtils';
 import { TargetValueThreeMonths } from '../TargetValueThreeMonths/TargetValueThreeMonths';
 import TransferRequestSection from './TransferRequestSection';
 import type { Meta } from '@storybook/react';
